Lab-IVR/Backend IVR/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.4.0
binaryornot==0.4.4
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
cryptography==46.0.5
deprecation==2.1.0
diagrams==0.24.4
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.9
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
mysql-connector-python==9.2.0
nodeenv==1.9.1
packaging==26.0
platformdirs==4.3.7
postgrest==2.28.0
pre_commit==4.2.0
propcache==0.4.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.1
pyiceberg==0.11.1
PyJWT==2.11.0
pyparsing==3.2.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.22
python-slugify==8.0.4
PyYAML==6.0.2
realtime==2.28.0
requests==2.32.3
rich==13.9.4
rsa==4.9
scapy==2.7.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
text-unidecode==1.3
tqdm==4.67.1
twilio==9.10.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.3.0
uvi==0.5.0
uvicorn==0.34.0
virtualenv==20.30.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.23.0
zstandard==0.25.0
